remove relationship between `InstalledApps` and `AppStorageStats` tables
